convert to TypeScript, https://github.com/phetsims/reactants-products-and-leftovers/issues/80

+++ js/common/view/StackNode.ts
@@ -1,8 +1,7 @@
 // Copyright 2014-2023, University of Colorado Boulder
 
-// @ts-nocheck
 /**
- * A vertical stack of Substances, built from the bottom up.
+ * StackNode is a vertical stack of Substances, built from the bottom up.
  * To improve performance:
  * - Nodes are created as needed.
  * - Nodes are never removed; they remain as children for the lifetime of this node.
@@ -11,40 +10,47 @@
  * @author Chris Malley (PixelZoom, Inc.)
  */
 
-import { Line, Node } from '../../../../scenery/js/imports.js';
+import TReadOnlyProperty from '../../../../axon/js/TReadOnlyProperty.js';
+import optionize, { EmptySelfOptions } from '../../../../phet-core/js/optionize.js';
+import { Line, Node, NodeOptions, NodeTranslationOptions } from '../../../../scenery/js/imports.js';
 import reactantsProductsAndLeftovers from '../../reactantsProductsAndLeftovers.js';
 import SubstanceIcon from './SubstanceIcon.js';
 
+type SelfOptions = EmptySelfOptions;
+
+type StackNodeOptions = SelfOptions & NodeTranslationOptions;
+
 export default class StackNode extends Node {
 
-  /**
-   * @param {number} height height of the stack
-   * @param {Property.<Node>} iconProperty the icon to display
-   * @param {Property.<number>} quantityProperty the number of nodes to display
-   * @param {number} startCenterY the centerY of the bottom node in the stack
-   * @param {number} deltaY the vertical spacing between nodes in the stack
-   * @param {Object} [options]
-   */
-  constructor( height, iconProperty, quantityProperty, startCenterY, deltaY, options ) {
+  private readonly disposeStackNode: () => void;
 
-    super();
+  /**
+   * @param height - height of the stack
+   * @param iconProperty - the icon to display
+   * @param quantityProperty - the number of nodes to display
+   * @param startCenterY - the centerY of the bottom node in the stack
+   * @param deltaY - the vertical spacing between nodes in the stack
+   * @param [providedOptions]
+   */
+  public constructor( height: number, iconProperty: TReadOnlyProperty<Node>, quantityProperty: TReadOnlyProperty<number>,
+                      startCenterY: number, deltaY: number, providedOptions?: StackNodeOptions ) {
+
+    const options = optionize<StackNodeOptions, SelfOptions, NodeOptions>()( {}, providedOptions );
 
     /*
      * This line is not visible (has no stroke), but defines the height of the stack,
      * and ensures that the stack always has well-defined bounds, even when the stack
      * contains zero items. All other nodes will be horizontally centered on this line.
      */
-    this.addChild( new Line( 0, 0, 0, height ) );
+    const line = new Line( 0, 0, 0, height );
 
     // parent for all items in the stack
     const itemsParent = new Node();
-    this.addChild( itemsParent );
 
-    // {SubstanceIcon[]}
-    const iconNodes = [];
+    const iconNodes: SubstanceIcon[] = [];
 
     // update the number of nodes in the stack
-    const quantityPropertyObserver = quantity => {
+    const quantityPropertyObserver = ( quantity: number ) => {
 
       const count = Math.max( quantity, itemsParent.getChildrenCount() );
 
@@ -66,21 +72,20 @@ export default class StackNode extends Node {
     };
     quantityProperty.link( quantityPropertyObserver ); // must be unlinked in dispose
 
-    this.mutate( options );
+    options.children = [ line, itemsParent ];
 
-    // @private
+    super( options );
+
     this.disposeStackNode = () => {
       iconNodes.forEach( node => node.dispose() );
       iconNodes.length = 0;
-      quantityProperty.unlink( quantityPropertyObserver );
+      if ( quantityProperty.hasListener( quantityPropertyObserver ) ) {
+        quantityProperty.unlink( quantityPropertyObserver );
+      }
     };
   }
 
-  /**
-   * @public
-   * @override
-   */
-  dispose() {
+  public override dispose(): void {
     this.disposeStackNode();
     super.dispose();
   }
